dynamic_mcp/requirements.txt
fastapi
uvicorn[standard]
sqlparse
google-genai
